Sundry list of issues fixed
Closes https://github.com/dalehenrich/Rowan/issues/79
If Rowan global or $ROWAN_HOME is not set, get a nice dialog
and logout.

The project list is alphabetical now.

Closes https://github.com/dalehenrich/Rowan/issues/110
Jadeite version number displayed in most of the title bars

Closes https://github.com/dalehenrich/Rowan/issues/106
The mystical 'dkh' timestamp is no longer displayed in the changes list.
No need for @dalehenrich to do anything.

Closes https://github.com/dalehenrich/Rowan/issues/67
Subclass creation on hierarchy class list shows correctly.
Add subclass definition is correct for the selected class.

Fixed unreported bug when doing an Add Subclass in the browser.

Closes https://github.com/dalehenrich/Rowan/issues/98
Project List Browser is aptly named.
Removed the Untitled in the Jadeite Login window.

Closes https://github.com/dalehenrich/Rowan/issues/86
If an .stb file is corrupted, don't throw an error. Just ignore
the walkback and let Jadeite open up a browser in the default
format.