Fix #63 and #64
A lot of refactoring to make #63 method list save work with the same code base as the JadeiteProjectsBrowserPresenter.

Smartened update code so if you have the same method open in two browser and one is changed it will update the other but only if there are no modifications pending. If there are mods in the other browser, "pink it".