Fixes #106. String inspector acting oddly.
If we encounter an oop type conversion error in an string execution, we throw an error now. And we do so outside the timing block. This should help prevent conditions where ad-hoc executions can start throwing up dialogs and hang the image.

Also wrap the jade inspect in an error block to give a little nicer message if the conditions described in #106 recur.